pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
dynamic = ["version"]

name = "spwn"
description = "Automatic tool to quickly start a pwn CTF challenge"
authors = [
	{name="Chino"},
	{name="Matteo Chiesa", email="matteo.chiesa.mc@gmail.com"}
]
maintainers = [
	{name="Matteo Chiesa", email="matteo.chiesa.mc@gmail.com"}
]

readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}

requires-python = ">=3.10"
dependencies = [
	"pwntools",
	"yara-python"
]

keywords = ["pwn", "ctf"]
classifiers = [
	"Environment :: Console",
	"Development Status :: 2 - Pre-Alpha",
	"License :: OSI Approved :: MIT License",
	"Intended Audience :: Developers",
	"Operating System :: POSIX :: Linux",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3",
]

[project.urls]
Homepage = "https://github.com/Church-17/spwn"

[project.scripts]
spwn = "spwn.main:main"

[tool.hatch.version]
path = "src/spwn/__init__.py"
